git-svn-id: https://jwutils.googlecode.com/svn/trunk@115 9038e2a8-f00e-11dd-89c5-9df526a60542

+++ projects/mobile/jmlogo/jmlogo/src/org/opengroove/jw/jmlogo/utils/Math2.java
@@ -26,6 +26,6 @@ public class Math2
      */
     public int ri(double value)
     {
-        return round(value);
+        return (int) round(value);
     }
 }
